+++ backend/core/src/jurisdictions/dto/jurisdiction.dto.ts
@@ -1,9 +1,10 @@
 import { OmitType } from "@nestjs/swagger"
 import { Expose, Type } from "class-transformer"
-import { ArrayMaxSize, IsArray, ValidateNested } from "class-validator"
+import { ArrayMaxSize, IsArray, IsString, ValidateNested } from "class-validator"
 import { ValidationsGroupsEnum } from "../../shared/types/validations-groups-enum"
 import { Jurisdiction } from "../entities/jurisdiction.entity"
 import { IdDto } from "../../shared/dto/id.dto"
+import { IdNameDto } from "../../shared/dto/idName.dto"
 
 export class JurisdictionDto extends OmitType(Jurisdiction, ["preferences", "programs"] as const) {
   @Expose()
@@ -19,4 +20,10 @@ export class JurisdictionDto extends OmitType(Jurisdiction, ["preferences", "pro
   @ValidateNested({ groups: [ValidationsGroupsEnum.default] })
   @Type(() => IdDto)
   preferences: IdDto[]
+}
+
+export class JurisdictionSlimDto extends IdNameDto {
+  @Expose()
+  @IsString({ groups: [ValidationsGroupsEnum.default] })
+  publicUrl: string
 }