+++ sites/partners/src/listings/Aside.tsx
@@ -1,6 +1,5 @@
-import React, { useContext, useMemo, useCallback } from "react"
+import React, { useContext, useMemo } from "react"
 import dayjs from "dayjs"
-import { useFormContext } from "react-hook-form"
 import {
   t,
   StatusAside,
@@ -15,7 +14,6 @@ import {
 } from "@bloom-housing/ui-components"
 import { pdfUrlFromListingEvents } from "@bloom-housing/shared-helpers"
 import { ListingContext } from "./ListingContext"
-import { createDate } from "../../lib/helpers"
 import { ListingEventType, ListingStatus } from "@bloom-housing/backend-core/types"
 
 type AsideProps = {
@@ -34,8 +32,6 @@ const Aside = ({
   submitFormWithStatus,
 }: AsideProps) => {
   const listing = useContext(ListingContext)
-  // eslint-disable-next-line @typescript-eslint/unbound-method
-  const { getValues } = useFormContext() || {}
 
   const listingId = listing?.id
 
@@ -205,7 +201,10 @@ const Aside = ({
     if (type === "details") {
       elements.push(
         <GridCell key="btn-preview">
-          <a target="_blank" href={`${process.env.publicBaseUrl}/preview/listings/${listingId}`}>
+          <a
+            target="_blank"
+            href={`${listing.jurisdiction.publicUrl}/preview/listings/${listingId}`}
+          >
             <Button fullWidth onClick={() => false} type="button">
               {t("listings.actions.preview")}
             </Button>