+++ sites/partners/src/listings/PaperListingForm/index.tsx
@@ -11,7 +11,6 @@ import {
   StatusBar,
   AppearanceStyleType,
   Button,
-  TimeFieldPeriod,
   Modal,
   AppearanceBorderType,
   Tabs,
@@ -23,24 +22,13 @@ import {
 import { useForm, FormProvider } from "react-hook-form"
 import {
   ListingStatus,
-  ListingApplicationAddressType,
-  Unit,
-  Listing,
-  ListingEvent,
   ListingEventType,
-  ListingEventCreate,
   Preference,
-  PaperApplication,
-  PaperApplicationCreate,
-  ListingReviewOrder,
-  User,
-  ListingPreference,
   Program,
-  ListingProgram,
 } from "@bloom-housing/backend-core/types"
-import { YesNoAnswer } from "../../applications/PaperApplicationForm/FormTypes"
-import dayjs from "dayjs"
 import { nanoid } from "nanoid"
+import { AlertErrorType, FormListing, TempEvent, TempUnit, formDefaults } from "./formTypes"
+import ListingDataPipeline from "./ListingDataPipeline"
 
 import Aside from "../Aside"
 import AdditionalDetails from "./sections/AdditionalDetails"
@@ -48,13 +36,6 @@ import AdditionalEligibility from "./sections/AdditionalEligibility"
 import LeasingAgent from "./sections/LeasingAgent"
 import AdditionalFees from "./sections/AdditionalFees"
 import Units from "./sections/Units"
-import {
-  stringToBoolean,
-  stringToNumber,
-  createDate,
-  createTime,
-  removeEmptyObjects,
-} from "../../../lib/helpers"
 import BuildingDetails from "./sections/BuildingDetails"
 import ListingIntro from "./sections/ListingIntro"
 import ListingPhoto from "./sections/ListingPhoto"
@@ -70,385 +51,14 @@ import BuildingSelectionCriteria from "./sections/BuildingSelectionCriteria"
 import { getReadableErrorMessage } from "../PaperListingDetails/sections/helpers"
 import { useJurisdictionalPreferenceList, useJurisdictionalProgramList } from "../../../lib/hooks"
 
-export type FormListing = Omit<Listing, "countyCode"> & {
-  applicationDueDateField?: {
-    month: string
-    day: string
-    year: string
-  }
-  applicationDueTimeField?: {
-    hours: string
-    minutes: string
-    period: TimeFieldPeriod
-  }
-  postmarkByDateDateField?: {
-    month: string
-    day: string
-    year: string
-  }
-  postmarkByDateTimeField?: {
-    hours: string
-    minutes: string
-    period: TimeFieldPeriod
-  }
-  arePaperAppsMailedToAnotherAddress?: YesNoAnswer
-  arePostmarksConsidered?: YesNoAnswer
-  canApplicationsBeDroppedOff?: YesNoAnswer
-  canPaperApplicationsBePickedUp?: YesNoAnswer
-  digitalApplicationChoice?: YesNoAnswer
-  commonDigitalApplicationChoice?: YesNoAnswer
-  paperApplicationChoice?: YesNoAnswer
-  referralOpportunityChoice?: YesNoAnswer
-  dueDateQuestionChoice?: YesNoAnswer
-  criteriaAttachType?: string
-  lotteryDate?: {
-    month: string
-    day: string
-    year: string
-  }
-  lotteryStartTime?: {
-    hours: string
-    minutes: string
-    period: TimeFieldPeriod
-  }
-  lotteryEndTime?: {
-    hours: string
-    minutes: string
-    period: TimeFieldPeriod
-  }
-  lotteryDateNotes?: string
-  postMarkDate?: {
-    month: string
-    day: string
-    year: string
-  }
-  reviewOrderQuestion?: string
-  waitlistOpenQuestion?: YesNoAnswer
-  waitlistSizeQuestion?: YesNoAnswer
-  whereApplicationsDroppedOff?: ListingApplicationAddressType
-  whereApplicationsPickedUp?: ListingApplicationAddressType
-}
-
-export const addressTypes = {
-  ...ListingApplicationAddressType,
-  anotherAddress: "anotherAddress",
-}
-
 type ListingFormProps = {
   listing?: FormListing
   editMode?: boolean
 }
 
-type AlertErrorType = "api" | "form"
-
-const defaults: FormListing = {
-  id: undefined,
-  createdAt: undefined,
-  updatedAt: undefined,
-  applicationDueDate: null,
-  applicationDueTime: null,
-  applicationFee: null,
-  applicationMethods: [],
-  applicationOpenDate: new Date(dayjs().subtract(10).format()),
-  applicationOrganization: "",
-  applicationPickUpAddress: null,
-  applicationPickUpAddressOfficeHours: "",
-  applicationMailingAddress: null,
-  applicationDropOffAddress: null,
-  applicationDropOffAddressOfficeHours: null,
-  assets: [],
-  buildingSelectionCriteria: "",
-  buildingSelectionCriteriaFile: null,
-  criteriaAttachType: "",
-  jurisdiction: undefined,
-  costsNotIncluded: "",
-  creditHistory: "",
-  criminalBackground: "",
-  depositMax: "0",
-  depositMin: "0",
-  depositHelperText: "or one month's rent may be higher for lower credit scores",
-  disableUnitsAccordion: false,
-  displayWaitlistSize: false,
-  events: [],
-  image: null,
-  leasingAgentAddress: null,
-  leasingAgentEmail: null,
-  leasingAgentName: null,
-  leasingAgentOfficeHours: "",
-  leasingAgentPhone: null,
-  leasingAgentTitle: "",
-  name: null,
-  postMarkDate: null,
-  postmarkedApplicationsReceivedByDate: null,
-  listingPreferences: [],
-  listingPrograms: [],
-  programRules: "",
-  rentalAssistance:
-    "The property is subsidized by the Section 8 Project-Based Voucher Program. As a result, Housing Choice Vouchers, Section 8 and other valid rental assistance programs are not accepted by this property.",
-  rentalHistory: "",
-  requiredDocuments: "",
-  status: ListingStatus.pending,
-  waitlistCurrentSize: null,
-  waitlistMaxSize: null,
-  isWaitlistOpen: null,
-  waitlistOpenSpots: null,
-  whatToExpect:
-    "Applicants will be contacted by the property agent in rank order until vacancies are filled. All of the information that you have provided will be verified and your eligibility confirmed. Your application will be removed from the waitlist if you have made any fraudulent statements. If we cannot verify a housing preference that you have claimed, you will not receive the preference but will not be otherwise penalized. Should your application be chosen, be prepared to fill out a more detailed application and provide required supporting documents.",
-  units: [],
-  accessibility: "",
-  amenities: "",
-  buildingAddress: null,
-  buildingTotalUnits: 0,
-  developer: null,
-  householdSizeMax: 0,
-  householdSizeMin: 0,
-  neighborhood: "",
-  petPolicy: "",
-  smokingPolicy: "",
-  unitsAvailable: 0,
-  unitAmenities: "",
-  servicesOffered: "",
-  yearBuilt: null,
-  urlSlug: undefined,
-  showWaitlist: false,
-  reviewOrderType: null,
-  unitsSummary: [],
-  unitsSummarized: {
-    unitTypes: [],
-    priorityTypes: [],
-    amiPercentages: [],
-    byUnitTypeAndRent: [],
-    byUnitType: [],
-    byAMI: [],
-    hmi: {
-      columns: [],
-      rows: [],
-    },
-  },
-}
-
-export type TempUnit = Unit & {
-  tempId?: number
-  maxIncomeHouseholdSize1?: string
-  maxIncomeHouseholdSize2?: string
-  maxIncomeHouseholdSize3?: string
-  maxIncomeHouseholdSize4?: string
-  maxIncomeHouseholdSize5?: string
-  maxIncomeHouseholdSize6?: string
-  maxIncomeHouseholdSize7?: string
-  maxIncomeHouseholdSize8?: string
-}
-
-export type TempEvent = ListingEvent & {
-  tempId?: string
-}
-
-export type PaperApplicationHybrid = PaperApplication | PaperApplicationCreate
-
-const formatFormData = (
-  data: FormListing,
-  units: TempUnit[],
-  openHouseEvents: TempEvent[],
-  preferences: ListingPreference[],
-  programs: ListingProgram[],
-  saveLatLong: LatitudeLongitude,
-  customPinPositionChosen: boolean,
-  profile: User
-) => {
-  const applicationDueDateFormatted = createDate(data.applicationDueDateField)
-  const applicationDueTimeFormatted = createTime(
-    applicationDueDateFormatted,
-    data.applicationDueTimeField
-  )
-
-  let postmarkByDateTimeFormatted = null
-
-  if (data.arePostmarksConsidered === YesNoAnswer.Yes && data.postmarkByDateDateField) {
-    const postmarkByDateFormatted = createDate(data.postmarkByDateDateField)
-    if (data.postmarkByDateTimeField?.hours) {
-      postmarkByDateTimeFormatted = createTime(
-        postmarkByDateFormatted,
-        data.postmarkByDateTimeField
-      )
-    } else {
-      postmarkByDateTimeFormatted = postmarkByDateFormatted
-    }
-  }
-
-  units.forEach((unit) => {
-    switch (unit.unitType?.name) {
-      case "fourBdrm":
-        unit.numBedrooms = 4
-        break
-      case "threeBdrm":
-        unit.numBedrooms = 3
-        break
-      case "twoBdrm":
-        unit.numBedrooms = 2
-        break
-      case "oneBdrm":
-        unit.numBedrooms = 1
-        break
-      default:
-        unit.numBedrooms = null
-    }
-
-    Object.keys(unit).forEach((key) => {
-      if (key.indexOf("maxIncomeHouseholdSize") >= 0) {
-        if (parseInt(unit[key])) {
-          if (!unit.amiChartOverride) {
-            unit.amiChartOverride = {
-              id: undefined,
-              createdAt: undefined,
-              updatedAt: undefined,
-              items: [],
-            }
-          }
-          unit.amiChartOverride.items.push({
-            percentOfAmi: parseInt(unit.amiPercentage),
-            householdSize: parseInt(key[key.length - 1]),
-            income: parseInt(unit[key]),
-          })
-        }
-      }
-    })
-
-    unit.floor = stringToNumber(unit.floor)
-    unit.maxOccupancy = stringToNumber(unit.maxOccupancy)
-    unit.minOccupancy = stringToNumber(unit.minOccupancy)
-    unit.numBathrooms = stringToNumber(unit.numBathrooms)
-
-    if (!unit.sqFeet) {
-      delete unit.sqFeet
-    }
-
-    delete unit.tempId
-  })
-
-  const events: ListingEventCreate[] = data.events?.filter(
-    (event) => !(event?.type === ListingEventType.publicLottery)
-  )
-  if (
-    data.lotteryDate &&
-    data.lotteryDate.day &&
-    data.lotteryDate.month &&
-    data.lotteryDate.year &&
-    data.reviewOrderQuestion === "reviewOrderLottery"
-  ) {
-    const startTime = createTime(createDate(data.lotteryDate), data.lotteryStartTime)
-    const endTime = createTime(createDate(data.lotteryDate), data.lotteryEndTime)
-
-    events.push({
-      type: ListingEventType.publicLottery,
-      startTime: startTime,
-      endTime: endTime,
-      note: data.lotteryDateNotes,
-    })
-  }
-
-  if (openHouseEvents) {
-    openHouseEvents.forEach((event) => {
-      events.push({
-        type: ListingEventType.openHouse,
-        ...event,
-      })
-    })
-  }
-
-  const jurisdiction =
-    !data.jurisdiction?.name && profile.jurisdictions.length === 1
-      ? profile.jurisdictions[0]
-      : data.jurisdiction
-
-  return {
-    ...data,
-    jurisdiction,
-    applicationDueTime: applicationDueTimeFormatted,
-    disableUnitsAccordion: stringToBoolean(data.disableUnitsAccordion),
-    units: units,
-    listingPreferences: preferences,
-    listingPrograms: programs,
-    buildingAddress: {
-      ...data.buildingAddress,
-      latitude: saveLatLong.latitude ?? null,
-      longitude: saveLatLong.longitude ?? null,
-    },
-    customMapPin: customPinPositionChosen,
-    isWaitlistOpen:
-      data.waitlistOpenQuestion === YesNoAnswer.Yes
-        ? true
-        : data.waitlistOpenQuestion === YesNoAnswer.No
-        ? false
-        : null,
-    applicationDueDate: applicationDueDateFormatted,
-    yearBuilt: data.yearBuilt ? Number(data.yearBuilt) : null,
-    waitlistCurrentSize:
-      data.waitlistCurrentSize && data.waitlistOpenQuestion === YesNoAnswer.Yes
-        ? Number(data.waitlistCurrentSize)
-        : null,
-    waitlistMaxSize:
-      data.waitlistMaxSize && data.waitlistOpenQuestion === YesNoAnswer.Yes
-        ? Number(data.waitlistMaxSize)
-        : null,
-    waitlistOpenSpots:
-      data.waitlistOpenSpots && data.waitlistOpenQuestion === YesNoAnswer.Yes
-        ? Number(data.waitlistOpenSpots)
-        : null,
-    postmarkedApplicationsReceivedByDate: postmarkByDateTimeFormatted,
-    applicationDropOffAddressType:
-      data.canApplicationsBeDroppedOff === YesNoAnswer.Yes &&
-      addressTypes[data.whereApplicationsDroppedOff] !== addressTypes.anotherAddress
-        ? addressTypes[data.whereApplicationsDroppedOff]
-        : null,
-    applicationPickUpAddressType:
-      data.canPaperApplicationsBePickedUp === YesNoAnswer.Yes &&
-      addressTypes[data.whereApplicationsPickedUp] !== addressTypes.anotherAddress
-        ? addressTypes[data.whereApplicationsPickedUp]
-        : null,
-    applicationDropOffAddress:
-      data.canApplicationsBeDroppedOff === YesNoAnswer.Yes &&
-      data.whereApplicationsDroppedOff === addressTypes.anotherAddress
-        ? data.applicationDropOffAddress
-        : null,
-    applicationPickUpAddress:
-      data.canPaperApplicationsBePickedUp === YesNoAnswer.Yes &&
-      data.whereApplicationsPickedUp === addressTypes.anotherAddress
-        ? data.applicationPickUpAddress
-        : null,
-    applicationMailingAddress: data.arePaperAppsMailedToAnotherAddress
-      ? data.applicationMailingAddress
-      : null,
-    events,
-    reservedCommunityType: data.reservedCommunityType.id ? data.reservedCommunityType : null,
-    reviewOrderType:
-      data.reviewOrderQuestion === "reviewOrderLottery"
-        ? ListingReviewOrder.lottery
-        : ListingReviewOrder.firstComeFirstServe,
-    digitalApplication:
-      data.digitalApplicationChoice === YesNoAnswer.Yes
-        ? true
-        : data.digitalApplicationChoice === YesNoAnswer.No
-        ? false
-        : null,
-    commonDigitalApplication: data.commonDigitalApplicationChoice === YesNoAnswer.Yes,
-    paperApplication:
-      data.paperApplicationChoice === YesNoAnswer.Yes
-        ? true
-        : data.paperApplicationChoice === YesNoAnswer.No
-        ? false
-        : null,
-    referralOpportunity:
-      data.referralOpportunityChoice === YesNoAnswer.Yes
-        ? true
-        : data.referralOpportunityChoice === YesNoAnswer.No
-        ? false
-        : null,
-  }
-}
-
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 const ListingForm = ({ listing, editMode }: ListingFormProps) => {
-  const defaultValues = editMode ? listing : defaults
+  const defaultValues = editMode ? listing : formDefaults
   const formMethods = useForm<FormListing>({
     defaultValues,
     shouldUnregister: false,
@@ -562,27 +172,22 @@ const ListingForm = ({ listing, editMode }: ListingFormProps) => {
         try {
           setLoading(true)
           clearErrors()
-          const orderedPreferences = preferences.map((preference, index) => {
-            return { preference, ordinal: index + 1 }
-          })
-          const orderedPrograms = programs.map((program, index) => {
-            return { program: { ...program }, ordinal: index + 1 }
-          })
-          const formattedData = formatFormData(
-            formData,
+
+          const dataPipeline = new ListingDataPipeline(formData, {
+            preferences,
+            programs,
             units,
             openHouseEvents,
-            orderedPreferences,
-            orderedPrograms,
+            profile,
             latLong,
             customMapPositionChosen,
-            profile
-          )
-          removeEmptyObjects(formattedData)
+          })
+          const formattedData = await dataPipeline.run()
+
           const result = editMode
             ? await listingsService.update(
                 {
-                  listingId: listing.id,
+                  id: listing.id,
                   body: { id: listing.id, ...formattedData },
                 },
                 { headers: { "x-purge-cache": true } }
@@ -592,6 +197,7 @@ const ListingForm = ({ listing, editMode }: ListingFormProps) => {
                 { headers: { "x-purge-cache": true } }
               )
           reset(formData)
+
           if (result) {
             /**
              * Send purge request to Nginx.